refacto: move `ZFPeriodDescriptor`, `PeriodicHLC` in their own files
