+++ zenoh-flow/src/model/mod.rs
@@ -16,3 +16,4 @@ pub mod connector;
 pub mod dataflow;
 pub mod link;
 pub mod operator;
+pub mod period;
