+++ zenoh-flow/src/runtime/runner.rs
@@ -16,8 +16,9 @@ use crate::runtime::connectors::{ZFZenohReceiver, ZFZenohSender};
 use crate::runtime::graph::link::{ZFLinkReceiver, ZFLinkSender};
 use crate::runtime::message::{Message, ZFMessage};
 use crate::types::{Token, ZFContext, ZFData, ZFInput, ZFResult};
-use crate::{async_std::sync::Arc, PeriodicHLC};
+use crate::utils::hlc::PeriodicHLC;
 use crate::{OperatorTrait, SinkTrait, SourceTrait};
+use async_std::sync::Arc;
 use futures::future;
 use libloading::Library;
 use std::collections::HashMap;
